+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [playground-quarkus] code review to remove bootstrap
 - [playground-quarkus] default html type handler set to FreeMarkerHtmlFragmentTypeHandlerEscapeUTF8
 
+### Fixed
+
+- [playground-quarkus] <#ftl> directive handling
+
 ## [3.1.7] - 2023-10-21
 
 ### Added